status: support JSON output format
status now supports to print the information in JSON format.
The format can be selected via the new parameter "--format", it accepts "plain",
"csv" and "json" as value.
In the JSON output fields that are not specified via "--fields" are omitted.
Fields that have no value, e.g. RunID when a task status is pending, is nil.

Because we want to distinguish between fields that should be omitted, fields
that undefined and empty values, we can not marshal a struct directly. It can
not distinguish between omitted fields and undefined fields.
To realize that the structs holding the output data is converted to a map and
then marshaled, fields that are omitted are not added to the map.

--format replaces the previous "--csv" parameters.
The "--csv" parameter for status is deprecated