formatter: align json implementation
Provide a JSON formatter that implements the same Formatter interface as the
other implentations.
The JSON formatter, accepts during intialization a list of field names.
When WriteRows() is called, it accepts the value to be in the field names
order. For every row it creates a Field-Name:Value map and appends it to an
internal slice.
The slice is marshalled and written to an io.Writer when Flush() is called.

This allows to converge the code for using the JSON formatter with the other
output formats.

The "AppName" field in the JSON output of "ls apps" is also changed to "Name".