inputresolver: fix: git objects for files in symlinked directory not found
Since "inputfiles: track symlink target paths" (e834583c), git objects for files
in a symlinked directory were not found anymore.

It was caused by multiple issues:
- it was only checked if the last component of the path was a symlinked, if
  another component was, it was missing to resolve the realpath and try another
  lookup in the gitTrackedDb
- When the gitTrackedDb was queried and the path not found, it was never checked
  if an entry for the Symlink target path existed because of an erroneous
  obj.Mode check, this is fixed

The Input.repoRelSymlinkTargetPath field and related methods and fields are
renamed to Input.repoRelRealPath, this naming also covers cases where any part
of a path component is a symlink.

Testcases to cover the functionality are added.