tests: fix TestRunSimultaneously flakiness
The test was sometimes failing on windows because the start time of one of the
tasks was 1 second earlier then the one one of taskruntimes[0].startTime.

All tasks run for >=3sec, relax the test condition to also accept start and end
times that are 1 second off.
This could e.g. happen when the task start in a different order and the clock is
close to the next full second.