resolve/gosource: add rel file query with multiple main packages test
Add a testcase for the gosource resolver that specifies as file query pattern
cmd/**/main.go and runs the goresolver from a different directory then the test
(aka app dir) directory.

This testcase currently fails because the relative file query patterns are not
relative to the specified working dir (app dir) but to the cwd
(https://github.com/golang/go/issues/65965).

The existing testcase query_main_file is refactored for this purpose
The test_config.json files got a new field WorkingDir, the $WORKDIR
test_config.json variable is renamed to $TESTDIR.