git: simplify detecting if baur repository is part of a git repository
IsGitDir() was checking if the passed directory contains a ".git" directory.
If not it invoked "git rev-parse" and returned true if it succeeded.

Make the check consistent to always checking if a .git directory exists in the
passed or a parent directory. This ensures the same result independent of the
passed directory.
This also allows the check to succeed if no commit exist in the repository.