inputs: add new input type TaskInfo
This commit adds the TaskInfo Input type.
A TaskInfo input can be defined via the following section in an .app.toml file:
  [[Task.Input.TaskInfos]]
    task_name = "check"
    env_var = "CHECK_TASK_INFO_FILEPATH"

When the task with the TaskInfo Input is run,
information of the <task_name> task of the same baur app are provided in a
temporary file. The information are JSON
encoded. The environment variable <env_var> is set to contain the path to the
temporary file when the task's command is run.
The TaskInfo file contains total input digest, directory path and outputs of the
referenced Task.

In this example, when the task is run the information of the "check" task are
queried, the information are written to a file that is passed via the
CHECK_TASK_INFO_FILEPATH environment variable to the task command.
The content of the file could be:
  {
    "TotalInputDigest": "sha384:118cb50a647884f442deb9cc796f2590c443b8d2220b42e4502ffbc6f103f11a7ad2b3f0b8684d19f22bf704e7f599ae",
    "AppDir": "/tmp/baur-filesrc-test3814551359/app",
    "Outputs": [
      {
        "URI": "/tmp/TestRun_TaskInfoFileContent2435369936/001/out"
      }
    ]
  }

Depending on the status of the dependent tasks, the listed output URIs might or
might not exist. If the task status is exist, the outputs are queried from the
database for the existing run. If the task has not been run yet, the output
information are from the definition of the task. If the output URI is
deterministic and the task with the same inputs is run, the outputs will be
created.
baur does not order the task depending on the TaskInfos. That means if the
referenced task of TaskInfo Input and the task itself is run, it is not
guaranteed that the Outputs in the TaskInfo file exist.

Task Info outputs show up in "baur ls ouputs" as "task: <TASK-NAME>", their
digest is the total input digest of the task.

Cycles in TaskInfo definitions are detected when the baur configs are loaded.
When a cycle is found, baur fails with an error.

To the database the tables input_task and task_run_task_input are added that
store information about TaskInfo inputs. The required schema version is
increased.

Limitations:
- The --input-lookup-str parameter can not be used together with TaskInfos. I
  haven't found a good implementation to support both so far.
- It is only supported to use task of the same app as TaskInfo Inputs for now.
- If the same task is referenced for TaskInfos multiple times, it's inputs and
  status is evaluated multiple times. The status of a TaskInfo task is also
  queried multiple times if is also executed in the same "baur run".
- Because the TaskStatusEvaluator instance and it's referenced InputResolver now
  still exists while tasks are run, the memory usage of "baur run" is increased.