run: fix: panic on some exec errors
When executing a task command failed and the error isn't an ExitCodeError,
ErrTaskRunSkipped and neither a ErrUntrackedGitFilesExist error a nil pointer
dereference panic happened.

In the stderr.Printf() call a nil result pointer was accessed.
Do not access result if err is != nil, it is always nil in this case.