add "release create" command
Introduce a "release create" sub-command.
It creates a named snapshot of the state of a list of tasks in EXIST state.
Optionally user-specified metadata (arbitrary data) can be stored with the
release. Metadata is read from a file and stored in the database.

Releases will provide a static view of baur runs, including their
outputs, without having to ensure that the environment must be the same
when the outputs have been created.
When "baur status" + "baur ls outputs" is used there can be various
factors that make the result hard to reproduce, e.g:
- outputs might have been overwritten for the same input files via
  "baur run --force",
- change of the used tooling (go list) or  different environment
  variables used as inputs can result in different total input digests

In follow up further commands are added to interact with releases.