cmd: add baur release show
add the command "baur release show", it prints information about a stored
release in JSON format.
The metadata of a release can either be written to a separate file by providing
the "--metadata" parameter or when not specified, included base64 encoded in the
printed JSON format.

When querying release information from postgresql the existing query functions
are used, that retrieve more information than actually used to print release
information.
This could be optimized in the future, if it turns out to be necessary.