loader: fail if app names are not unique
Having multiple apps with the same app name is error prone.
When configuration files are loaded and the specifier is not
app-specific, fail if multiple app configs are found that use the same
app name value.

When loader specifiers are passed that are specific for an app like a
task id or "<APP-NAME>.*" multiple app configs defining the same app
name are not found. This is because the loader aborts the search as soon
as the wanted app is found.