init db: fix: specifying env db url requires repo config file
Specifying the postgresql url via the BAUR_POSTGRESQL_URL environment
variable only works when a baur repository config file exist.
When the repository config file is not found, it fails with an error.

This is not intentional, skip searching and reading the repository
config if the BAUR_POSTGRESQL_URL environment variable is set.