upgrade db: support defining min. and max. database schema version
Instead of requiring an up2date database schema allow to define a
minimum and maximum compatible database schema version.
The max. schema version is always the newest existing one at the time of
a release. The minimum schema version can be an older one if it is still
compatible.

This can make upgrading baur easier, when db schema changes are
downwards compatible, multiple baur version can be used with the same
database schema.
Even better would be to use a semantic version as schema version.
This would require more work though :-)