run: add --fail-fast parameter
Add a "--fail-fast" parameter to "baur run".
By default it is disabled, to keep downwards compatibility with the
current baur version

If enabled, "baur run" continues to execute tasks when one fails.
When disabled and a task execution to fails, execution of other queued
tasks are skipped. Concurrently running tasks are not aborted.