upgrade configs: succeed with a clear message when cfgs are up to date (#628)
When "baur upgrade configs" is run for configuration files that are in the
latest version, the command fails with an error message stating that configs can
only be upgraded from version 5 or 6.
Failing when everything is fine is confusing.

Succeed instead and print that the configs are already up to date and
nothings needs to be done.