resolve/gosource: remove files in GOMODCACHE from discovered inputs (#639)
This fixes that when the Gosource input resolver is used on a Go
application that uses non-vendored 3. party packages, baur fails with:

  ERROR: [..] evaluating task status failed: Rel: can't make
  $GOMODCACHE/[..] relative to /home/[...]

Imported 3. party packages that aren't vendored are stored by Go in the
GOMODCACHE directory.
When the Gosource resolver resolves a task to sources in the GOMODCACHE,
it replaces the path to the GOMODCACHE in the input file path with
'$GOMODCACHE'. This is done to always have the same reproducible input
files paths, independent of what the actual GOMODCACHE directory is.

The inputresolver afterwards tries to evaluate the repository relative
path of the input file containing the '$GOMODCACHE' which fails with the
error.
The real path to the input file is not known by the inputresolver, which
makes it difficult to track changes in the GOMODCACHE.

Instead of replacing the GOMODCACHE directory in paths, omit returning
files in the GOMODCACHE, same procedure as it is done for stdlib input files.

Instead a task can track and rely on the information in the go.mod or
go.sum files to track changes of 3. party libraries.
This has the disadvantage that if files in the GOMODCACHE are modified,
it won't be noticed by baur.