gosource: return path to go.mod files (#647)
The gosource resolver now returns the path to the go.mod as part of
the result.
The go.mod file defines, among others the imported third-party modules and
parameters for the build environment.

Including it in the gosource resolver results means, user do not need
define the go.mod file manually as part of baur tasks anymore.